Agfa35/LiveStream/requirements.txt
pyserial>=3.5
Pillow>=9.0.0
numpy>=1.21.0
